+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 
 * Added `nullptr`.
 
+* You can now take a pointer to a stack variable.
+
 * Arrays can now have elements of types other than `byte` (still limited in size to 1 byte though) and be built out of struct literals.
 
 * Arrays can now be constant.
@@ -63,6 +65,8 @@
 * Fixed `@long` and `@long_be` array filters.
 
 * 8080 and LR35902: fixed large stack variables.
+
+* Other bug fixes.
 
 * Optimization improvements.